models/wave2vec/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
huggingface-hub==0.17.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.2
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
pytorch-lightning==1.6.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sklearn==0.0.post9
soundfile==0.12.1
soxr==0.3.6
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.1.2
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2